app/requirements.txt
PyYAML==5.1.2
Flask==1.1.1
flask-mongoengine==0.9.5